Fix missing INFO log messages
This fixes an issue where logging.warning was called before the logging
was setup. This causes Python to call logging.basicConfig with the
defaults and hides INFO messages. It also overrides any future calls to
basicConfig.

This change just adds a call to the setup function for the early warning
logs.