gateware.usb: fix pending interrupts for USBDeviceController and eptri fifo peripherals
Interrupts were being triggered correctly for these peripherals but
the CSR pending register failed to show entries for the pending
peripheral, instead returning a zero mask.

This was due to upstream changes in lambdasoc EventSource which causes
event definitions to be created by default with a mode of "level".

There are three different supported modes:

    level: notification is raised when stb is high
    rise:  notification is raised on rising edge of stb
    fall:  notification is raised on falling edge of stb

Setting the mode to "rise" on events for these peripherals resolves
the problem.

Other usages of EventSource in our gateware already specify the mode
explicitly.