gateware.deps: update lambdasoc dependency specification
This PR primarily updates the lambdasoc dependency refence to point at
@mndza's fork of the repo:

    lambdasoc = {git = "https://github.com/mndza/lambdasoc.git"}

This is an up-to-date fork from the main lambdasoc repo that fixes a
regression that referenced an incorrect version of the litex bios
submodule.

Once https://github.com/lambdaconcept/lambdasoc/pull/18 is merged we
should be able to reference the main lambdasoc repo again.

There is also a small workaround for a bug where the litex bios
compilation directory is not created before a build.

The tracking issue for this is:

https://github.com/greatscottgadgets/luna/issues/191