gateware.architecture.adv: add ApolloAdvertiser
This module uses the FPGA_INT line present in recent Cynthion boards as
a one-way communication channel. Messages or "advertisements" are sent
periodically.

In this commit, it will announce its intention to acquire the main USB
port (CONTROL). Apollo can choose to honor this request and keep the
port handed off to the FPGA until these messages stop, which is
accomplished by asserting the `stop` signal or using the included
`ApolloAdvertiserRequestHandler` example.