gateware.stream.generator: StreamSerializer now counts the number of bytes sent
We should be able to use a `start_position` value greater than `max_length`,
which was not the case before this patch.

Typically when using this module to stream data to a USB control endpoint,
(in a vendor or class specific request handler), the `max_length` is set to 64
(this is the max packet size for control endpoint 0), and the `start_position`
is offset (0, 64, 128, etc) after each IN packet. This allows us to send
descriptors with >64B of content.